Added created_date for all models
There is a slight issue with this when using sqlite. See SO question:
http://stackoverflow.com/questions/3170634/how-to-solve-cannot-add-a-not
-null-column-with-default-value-null-in-sqlite3

Created-date should not be null, but when migrating to an existing
database, sqlite will error out if it finds a null column. Rather than
hack around this, I’m just going to blow away the db and alter my
pave_db workflow to make it work.

Created date can be null (for now).