Duplicate email and invalid user_id session state hack
This is a hack to prevent:

1. Duplicate emails getting added to database.
2. Receiving a valid login state containing a user_id of None.

I’m not terribly proud of this. Validation of duplicate emails should
probably be handled (in part) by the model.

A login session should always have a valid user_id assigned to it. The
short-circuiting here probably prevents that, but there should probably
protection via a throwable error somewhere.

Deals with #15